StaticAPI is functional
