+++ metra/__init__.py
@@ -1,18 +1,11 @@
+from .static import StaticAPI
 from .static import stops
-from .static import stop_times
 from .static import trips
 from .static import shapes
 from .static import routes
 from .static import calendar
 from .static import calendar_dates
 
-from .static import get_active_trips
-from .static import get_active_stop_times
-from .static import get_upcoming_stop_times
-from .static import get_route_stop_times
-from .static import update_stop_times
-
-from .utils import StaticAPI
 from .utils import get_publish_time
 from .utils import get_last_local_publish
 from .utils import update_schedule_zip
